fixed trash button

+++ src/components/CustomTemplates.vue
@@ -10,14 +10,15 @@
           <li v-for="file in files" :key="file" class="text-left text-xs font-mono pl-2 pr-1 md:px-4 md:pl-4 cursor-pointer py-1 border border-white hover:underline">
             <fa-icon :icon="['fas', 'file']" color="aqua" class="self-center mr-1"/>
             <span @click="downloadFile">{{file.split('/').pop()}}</span>
-            <button class="block mt-2 outline-none">
-              <fa-icon :icon="['fas', 'trash-alt']" class="text-xs" size="1x" color="red" :id="file.split('/').pop()" @click="deleteFile"/>
+            <br/>
+            <button class="outline-none w-12 h-8 z-40 text-red-600 hover:bg-red-600 hover:text-white rounded" :id="file.split('/').pop()" @click="deleteFile">
+              delete
             </button>
           </li>
         </ul>
       </div>
       <div>
-        <p class="text-xs text-center text-red-500">file preview(pdf only)</p>
+        <p class="text-xs text-center text-red-500">file preview(pdf,images,video)</p>
         <iframe src="" id="preview" class="w-full lg:h-64 mx-auto my-4 lg:max-w-sm"/>
         <MultipleFilesUploader category="Custom"/>
         <button class="btn w-full md:w-3/4 p-1 rounded my-3 bg-green-500 text-white hover:bg-green-700">
@@ -67,21 +68,28 @@ export default {
       }
     },
     async deleteFile(e){
+      let delFile = e.target.attributes.id.nodeValue
+      if(delFile){
       const del = window.confirm('Are you sure you want to delete this file?')
-      if(del){
-        this.spinner=true
-        let file = this.directoryPath.concat(e.target.attributes.id.nodeValue)
-        const storageRef=storage.ref()
-        let fileRef=storageRef.child(file.trim())
-        try{
-          await fileRef.delete()
-          this.spinner=false
-          window.alert(`${file} deleted Successfully`)
-        }catch(err){
-          this.spinner=false
-          window.alert(err.code)
-          console.log(err)
+        if(del){
+          this.spinner=true
+          console.log(`node value : ${delFile}`)
+          let file = this.directoryPath.concat(delFile)
+            const storageRef=storage.ref()
+            let fileRef=storageRef.child(file.trim())
+            try{
+              await fileRef.delete()
+              this.spinner=false
+              window.alert(`${file} deleted Successfully`)
+            }catch(err){
+              this.spinner=false
+              window.alert(err.code)
+              console.log(err)
+            }
         }
+      }else{
+        this.spinner=false
+        window.alert(`reference undefined, please try again`)
       }
     }
   },
